refactor: move positionStyles for sharing
