+++ src/client/features/Member/components/PositionBadge/index.tsx
@@ -5,7 +5,7 @@ import { PositionType } from 'server/actors/Members/constants/position';
 import { useAppTheme } from 'client/styles/tokens';
 import { componentElevationKey } from 'client/styles/elevation';
 import { Typography } from 'client/components/atoms/Typography';
-import { POSITION_STYLES } from 'client/features/Member/constants/positionStyles';
+import { POSITION_STYLES } from 'client/styles/positionStyles';
 
 const NoneBadge: React.FC = () => {
   const theme = useAppTheme();
