+++ src/client/features/Member/components/PositionCounter/index.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { useAppTheme } from 'client/styles/tokens';
 import { Typography } from 'client/components/atoms/Typography';
 import { componentElevationKey } from 'client/styles/elevation';
-import { POSITION_STYLES } from 'client/features/Member/constants/positionStyles';
+import { POSITION_STYLES } from 'client/styles/positionStyles';
 
 export type PositionCounterProps = {
   center: number;
